Removed comparisons of signed and unsigned integers

+++ dtcInterfaceLib/test/packetGenerator.cc
@@ -213,7 +213,7 @@ int main(int argc, char** argv)
 			inputCalDigiStream >> curHit.recoDigiT0;
 			inputCalDigiStream >> curHit.recoDigiSamples;
 			double curSample;
-			for(size_t curSampleNum = 0; curSampleNum < curHit.recoDigiSamples; curSampleNum++)
+			for(int curSampleNum = 0; curSampleNum < curHit.recoDigiSamples; curSampleNum++)
 			{
 			        inputCalDigiStream >> curSample;
 				curHit.waveform.push_back(curSample);
@@ -400,7 +400,7 @@ int main(int argc, char** argv)
 				// Find all hits for this event coming from the specified Ring/ROC
 				for(size_t curHitIdx=0; curHitIdx<calEventVector[eventNum].size(); curHitIdx++)
 				{
-				        if(calEventVector[eventNum][curHitIdx].rocID == rocID && calEventVector[eventNum][curHitIdx].ringID == ringID)
+				        if(calEventVector[eventNum][curHitIdx].rocID == (int)rocID && calEventVector[eventNum][curHitIdx].ringID == (int)ringID)
 					{
 					        curHitVector.push_back(calEventVector[eventNum][curHitIdx]);
 					}
@@ -489,7 +489,7 @@ int main(int argc, char** argv)
 						packetVector.push_back(IDNum);
 						packetVector.push_back((adc_t)(curHit.recoDigiT0));
 						packetVector.push_back((adc_t)(curHit.recoDigiSamples));
-						for (size_t sampleIdx = 0; sampleIdx < curHit.recoDigiSamples; sampleIdx++)
+						for (int sampleIdx = 0; sampleIdx < curHit.recoDigiSamples; sampleIdx++)
 						{
 						        adc_t scaledVal = curHit.waveform[sampleIdx];
 							packetVector.push_back(scaledVal);
